enhance AllocationListener with onRelease() and onKeepReleasedMemory() call
onRelease tells the caller that a memory will be released, it should return
true if the caller want to keep this memory, otherwise it should return false
and the memory will be released.
If onRelease return true, onKeepReleasedMemory will give the new arrow buffer
holding the memory to the caller, and it should return true if the caller
keep the new arrow buffer, otherwise it should return false and the memory
will be released.