ARROW-12154: [C++][Gandiva] Fix gandiva crash in certain OS/CPU combinations
Currently when running gandiva in a OS where the OS doesn't support all the features of the host cpu, specifically vector instructions like AVX, AVX512 which needs OS support (because the OS or the VM running the OS  is older version and doesn't support them, or passthrough is disabled for these features in the VM), llvm::sys::getHostCPUName detects the processor with these features and so gandiva generates jit compiled code with these vector instructions, which the cpu is unable to execute and hence faults.
Hence we need to detect the cpu features also and set it to the engine builder, since the OS might not support all the features of the detected cpu. This would also protect against other issues such as low end versions of SandyBridge, Haswell, and SkyLake processesors not supporting AVX at all.

Closes #9852 from projjal/llvmgetcpufeatures and squashes the following commits:

fef0fe8ec <Projjal Chanda> log cpu features
c909488cb <Projjal Chanda> Add cpu attributes to engine builder

Authored-by: Projjal Chanda <iam@pchanda.com>
Signed-off-by: Praveen <praveen@dremio.com>